!!serial: Make nulldev a multi-serial device
If CONFIG_SERIAL_MULTI and CONFIG_SYS_DEVICE_NULLDEV are defined, make
the nulldev a serial device instead of a console device.  This allows
the default console to be specified as the nulldev.  This is
specifically helpful when the real serial console's init() cannot run
early in the boot process.  When the init can be run, then the console
can be switched to the real device.

Signed-off-by: Joe Hershberger <joe.hershberger@ni.com>