nimyrio: Add initial support for myRIO targets
Add a packaged and board-only variant with unique DeviceCodes

Need to unlock NAND on myRIO

Manufacturing will likely use a USB Enet adapter (similar to Panda) for
talking to the myRIO.  This means they will need networking commands.

Also include the defethaddr for now so that we can generate on for the
Eth gadget in Linux (until mfg pulls from Oracle).

Signed-off-by: Joe Hershberger <joe.hershberger@ni.com>