nizynq: Prevent fdt from crossing 1MB memory boundary
It seems that if the FDT in physical memory crosses a 1MB boundary then
Linux will hang very early in the boot process.  This can happen at
random if a ramdisk is used because it is allocated first.

Work around the issue by forcing the ramdisk to start 32KB lower in
memory so that the fdt has a window available that is guaranteed
to not cross a 1MB boundary.

Signed-off-by: Joe Hershberger <joe.hershberger@ni.com>