nizynq: Fix FPGA loading scripts
These scripts were not using properly escaped quoting to allow for the
"test" command to work properly on empty / undefined variables.

We need to make sure all sources for setting the bootmode variable are
evaluated before the fpgaloadcmd is run.

Use a separate variable to track the success of loading an FPGA image
and go to recovery-mode if it was not successful.

Signed-off-by: Joe Hershberger <joe.hershberger@ni.com>
Acked-by: Jeff Westfahl <jeff.westfahl@ni.com>
Acked-by: Brad Mouring <brad.mouring@ni.com>