!!env: Handle write-once ethaddr and serial# generically
Use the variable access flags to implement the protection for ethaddr
and serial# instead of hard-coding them.

Series-to: u-boot
Series-cc: wd, ti, marex, Gerlando Falauto <gerlando.falauto@keymile.com>
Series-version: 3
Cover-letter:
Add environment call-back and flags capability

When a variable with a registered callback is inserted, deleted, or
overwritten the callback is called and gives the system an opportunity
to do something in response to the change.  It also has the opportunuty
to reject the change by returning non-zero.

The flags on variables can control their type as well as their allowed
access.

        The format of the list is:
                type_attribute = [s|d|x|b|i|m]
                attributes = type_attribute
                entry = variable_name[:attributes]
                list = entry[,list]

        The type attributes are:
                s - String (default)
                d - Decimal
                x - Hexadecimal
                b - Boolean ([1yYtT|0nNfF])
                i - IP address
                m - MAC address

        The access attributes are:
                a - Any (default)
                r - Read-only
                o - Write-once
                c - Change-default
END

Signed-off-by: Joe Hershberger <joe.hershberger@ni.com>