nizynq: Use env vars for backup addresses
MFG requested that these addresses be something they could read so that
their test need not change in the event that we have to move something
again.

Push these to the env where they can find them.  When we determine that
we need to restore, read the addresses from the env as well (default to
CPP as before).

Signed-off-by: Joe Hershberger <joe.hershberger@ni.com>