hush: Fix for and while loop regression
In commit b7a26e278a0e98043f460e1bac0202741d20b577 a regression was
introduced that caused while loops and for loops to only execute their
first iteration.  It was caused by the re-use of an "empty" pipe.
Unfortunately, the "done" at the end of a loop is also "empty" but it
has an r_mode that is set to RES_DONE instead of RES_NONE.  This means
that the done at the end of the loop was replaced by the following
command, thus preventing any looping from occurring.

We now preserve the (otherwise) unused pipe only when the context is
"DONE".  This termination keyword is used to acomplish looping.  I
checked the only other termination keyword "FI" and it does not seem
the care about sticking around.  It's sole responsibility is to pop
the "IF", "ELIF", or "ELSE" off of the stack.

Signed-off-by: Joe Hershberger <joe.hershberger@ni.com>
Acked-by: Jeff Westfahl <jeff.westfahl@ni.com>
Acked-by: Josh Cartwright <josh.cartwright@ni.com>