Xilinx: ARM: net: stop looking at rx_status in recv
In order for the driver to determine if there are more packets to
receive, first queries the framerx bit, if it is set, then it walks the
descriptor ring processing packets.  Once all the packets are received,
it writes the framerx bit back to the status register to clear the
condition.

However, the conditions under which this bit is set is unclear.  The
behavior we see is that the driver walks the descriptor ring once,
receiving 8 packets (u-boot configures the ring to be 8 buffers deep),
then stalls entirely until another frame is seen on the wire, at which
case the framerx bit is set again, causing the driver to go through the
ring again.

I say 'stalls' because in the case where we've accumulated packets,
there should be more packets that could be processed at that point; what
isn't clear is whether or not those packets have actually been written
to host memory.

So, I'm thinking that either:

  1. There is a race in u-boot driver, if the MAC DMAs a packet to host
     memory after the driver has walked the descriptor ring, but before
     the driver clears the framerx condition by writing to rxstatus

     If this is hit, then there would be packets in host memory that
     could be processed but aren't.

  2. The MAC is DMAing into the host buffers, but is not setting framerx
     until a subsequent frame is received.

  3. The MAC is not DMAing packets held in the hardware FIFO until it
     sees a new frame on the wire.

Number 3 seems like it should unlikely.  Regardless of which of these
cases we're seeing, the solution would be modifying the u-boot network
driver to not even look at/manage the framerx bit at all, and instead
unconditionally walk the descriptor ring.

It turns out, this is exactly what the more up-to-date u-boot zynq gem
driver is doing, complete with ominous comment:

	/* Do not check frame_recd flag in rx_status register 0x20 - just poll BD */

Signed-off-by: Josh Cartwright <josh.cartwright@ni.com>
Acked-by: Joe Hershberger <joe.hershberger@ni.com>
Acked-by: Jeff Westfahl <jeff.westfahl@ni.com>