nizynq: Lock all of the u-boot MTD partition
The blocks reserved for bad blocks in the u-boot partition must be
locked in addition to the ideal blocks.  In case someone mucks with
the MTD partition info in the u-boot env in such a way that the look-up
fails, still lock at least the ideal blocks.

The MTD info cannot be initialized inside of board_nand_init() because
the mtdparts_init() looks up the nand part which is in the middle of
being initialized and is not registered with MTD yet.  Move the
operation to board_late_init().

Signed-off-by: Joe Hershberger <joe.hershberger@ni.com>
Acked-by: Jeff Westfahl <jeff.westfahl@ni.com>
Acked-by: Josh Cartwright <josh.cartwright@ni.com>