part_dos: Actually check for valid partitions in MBR
Some "makeboot" tools will put "FAT32" in the MBR even though it's not
a "super floppy".  This confused u-boot into thinking it is a PBR.

Check for a valid boot indicator field in each partition to detect an
MBR instead of checking for the FAT strings.  This is closer to what
Linux does to choose between the two posibilities.

Signed-off-by: Joe Hershberger <joe.hershberger@ni.com>