nand: Reset after an operation fails due to lock
When a page is locked and an operation is attempted on it, the status
that is returned indicates "write protected".  This status will be
returned from then on until the state is reset.  Without the reset, the
NAND subsystem then assumes that the whole chip is write protected thus
preventing all non-read operations including unlock!

Reset the NAND state any time an operation completes with a status that
indicates "write protected".

Signed-off-by: Joe Hershberger <joe.hershberger@ni.com>