nicrio9068: Explicitly disable dcache when starting netconsole
On cRIO-9068, netconsole uses the Zynq built-in Ethernet MAC, whose driver
doesn't work with the dcache on. We should explicitly disable dcache when
we know we're going to be using that driver. This is in preparation for
turning on the dcache automatically at an early point in the boot for
performance reasons.

Signed-off-by: Jeff Westfahl <jeff.westfahl@ni.com>