nand: bbt: add a config option to not mark bbt blocks as bad
Normally, the BBT marks its own blocks as bad so that they can't be erased
except by the BBT itself. Unfortunately, this prevents us from using the
BBT to determine if its blocks are actually bad. Added a new config option
CONFIG_NAND_BBT_NO_MARK that prevents the BBT from marking its own blocks
as bad. This is dangerous as it allows the BBT blocks to be erased, but it
allows us to tell if the BBT blocks are actually bad using "nand bad".

Signed-off-by: Jeff Westfahl <jeff.westfahl@ni.com>