nizynq: mfg: don't mark BBT blocks
Normally, the BBT marks its own blocks as bad so that they can't be erased
except by the BBT itself. Unfortunately, this prevents us from using the
BBT to determine if its blocks are actually bad. For manufacturing, use
the config option CONFIG_NAND_BBT_NO_MARK so that the BBT won't mark its
own blocks as bad. This allows us to tell if the BBT blocks are actually
bad and take appropriate action.

Signed-off-by: Jeff Westfahl <jeff.westfahl@ni.com>