nimyrio: Use EEPROM MAC address in mfg
When manufacturing programs a board and boots from the u-boot on the
NAND, the MAC address is not yet configured.  If we detect this case
(because the backup area is also blank) we use the MAC address
configured in the USB dongle so that manufacturing can talk to the
board over NetConsole before the real MAC address is configured.

Signed-off-by: Joe Hershberger <joe.hershberger@ni.com>