nimyrio: Include a default NVS and preserve
Set a default NVS incase it is ever clobbered, the radio will still
function, but not be calibrated and not have the correct MAC address.

The eth1addr should also contain the wireless MAC address, so an
enterprising script could actually calibrate and recreate the wl12xxnvs
variable in u-boot.

We also need to make sure that we preserve the value of the nvs across
env reset and mark it as write-once in the env.

Signed-off-by: Joe Hershberger <joe.hershberger@ni.com>