Xilinx: ARM: i2c: Add a bus reset function
Added an I2C bus reset function for Zynq. This function changes the I2C
MIO pin configuration to GPIO, and checks if an I2C slave is driving SDA
low. This can happen if the Zynq was reset while reading a 0 from an I2C
slave. If SDA is low, we toggle SCL for up to 9 clock cycles, until the
I2C slave stops driving SDA low.

Signed-off-by: Jeff Westfahl <jeff.westfahl@ni.com>