Add ability to ignore environment and use defaults
At times when the saved environment is valid (i.e. the CRC is OK), but
non-functional, it is convenient to be able to still boot. We create
an option, CONFIG_FORCE_DEFAULT_ENV which, when enabled, will result
in a u-boot that will ignore the saved environment and use the default
environment. No attempt will be made at reading the saved environment
whatsoever.

To implement this functionality, we make all existing definitions of
env_relocate_spec weak and optionally, as according to
CONFIG_FORCE_DEFAULT_ENV, define a strong env_relocate_spec that
forces a default environment to override any already defined
env_relocate_spec.

Signed-off-by: Jaeden Amero <jaeden.amero@ni.com>