nisbrio9651: Force default environment in MFG mode
In manufacturing mode, we'd prefer to avoid reading any existing
environment. We expect any existing environment to be invalid, anyway.
This might end up speeding up boot times since a CRC won't be
calculated. This allows us to put the device into a known good state
if the environments are set to a state that causes the device to be
non-functional.

Signed-off-by: Jaeden Amero <jaeden.amero@ni.com>