nicrio9068: Properly control the RGMII logic
When we connect at 1000 Mbps, we need to set the link speed GPIO to 0
to tell the MAC to PHY logic in the Zynq PL to drive data on the
rising edge of the clock and different data on the falling edge of the
clock (RGMII at 1000 Mbps). When we connect at 10/100 Mbps, we need to
set the link speed GPIO to 1 to tell the MAC to PHY logic in the PL to
drive the same data on both the rising edge and falling edge of the
clock (RGMII at 10/100 Mbps).

Note that we previously were not setting the link speed GPIO and
things still seemed to work fine at 100 Mbps. This is due to the
specific PHY we are using on cRIO-9068 and could break when a
different PHY is used.

Signed-off-by: Jaeden Amero <jaeden.amero@ni.com>