zynq_gem: Don't output invalid PHY frame on reset
Previously, during XEmacPss_Reset, we were writing 0x00 to the phy_maint
(XEMACPS_PHYMNTNC) register in an attempt to set that register to its
power on reset and "Clear phy (if there is any previously detected)
address". However, writing to this register has a side effect, in that
whenever it is written to, its contents are shifted out over MDIO. In
order to produce a valid frame, bits 29 and 28 must be set to either
0b10 or 0b01 in order to indicate a read or write frame, and bits 17 and
16 must be set to 0b10 in order for the frame to be valid at all. By
shifting out the power on reset value, 0x00, we were emitting an invalid
PHY maintenance frame.

This comment removes the write to the phy_maint register. As part of
resetting the MAC, we don't need to write to the phy_maint register. We
certainly don't need to emit an invalid frame. Higher level software
will be responsible for configuring the PHY at a later time.

Signed-off-by: Jaeden Amero <jaeden.amero@ni.com>