nisbrio9651: Switch USB0 between host and device
Swtich USB0 to host if the USB OTGSC_ID bit is not set. Don't switch
USB0 to host if USB OTGSC_ID bit is set. When the provided device tree
has USB0's "dr_mode" set to "peripheral", this has the effect of making
USB0 host when the ID line is cleared and making it a peripheral when
the ID line is set.

Signed-off-by: Jaeden Amero <jaeden.amero@ni.com>
Acked-by: Jeff Westfahl <jeff.westfahl@ni.com>
Acked-by: Joe Hershberger <joe.hershberger@ni.com>