nisbrio9651: Use the standard usbgadgetethaddr name
We'd like to be able to use one set of kernel arguments for all of our
controllers in order to set the USB Gadget Ethernet address. This task
is made easiest when all of our controllers use a consistent name of
"usbgadgetethaddr" for the environment variable containing this address.
This commit updates the name from "eth2addr" to the standard
"usbgadgetethaddr" for sbRIO-9651 controllers.

Signed-off-by: Jaeden Amero <jaeden.amero@ni.com>