nimyrio: Use the standard usbgadgetethaddr name
We'd like to be able to use one set of kernel arguments for all of our
controllers in order to set the USB Gadget Ethernet address. This task
is made easiest when all of our controllers use a consistent name of
"usbgadgetethaddr" for the environment variable containing this address.
This commit adds the standard "usbgadgetethaddr" variable for use by
roboRIO and future controllers that are similar to myRIO. Older
controllers, like myRIO, have to keep using the old names for backwards
compatibility.

Signed-off-by: Jaeden Amero <jaeden.amero@ni.com>
Acked-by: Joe Hershberger <joe.hershberger@ni.com>
Acked-by: Keng Soon Cheah <keng.soon.cheah@ni.com>