zynq_gem: Implement the Zynq PS GEM halt function
The Xgmac_halt() function is blank. This could cause issue then
the system do not expect the ethernet controller to function. For
example, if the ethernet controller is not halted when the OS boots,
the ethernet controller could be writing to the memory region owned
by the OS and hence corrupting the memory content. Kernel crash had
been observed in real life example.

Implement the Xgmac_halt() by disabling the TX and RX.

Signed-off-by: Keng Soon Cheah <keng.soon.cheah@ni.com>