nisbrio9651: Add backup CCA serial offset
Manufacturing wants to test 4 sbRIO-9651 units at once, and to be able
to make sure that when they think they are testing sbRIO-9651 #1, they
are actually testing sbRIO-9651 #1 and not #2, #3, or #4. Manufacturing
plans on adding some safeguards to the test fixture to prevent such
mistakes, but they would like an additional level of protection.

Manufacturing would like to be able to program the CCA serial number
(S/N) into the backup area in the raw NAND and in an environment
variable. (Manufacturing already does the same thing for the Module
Assembly (MA) serial number.)

During the CCA-level test, the CCA S/N would get written to the backup
section of NAND.

During the MA-level test, the test would query Oracle for the child CCA
S/N. The test would then read the CCA S/N from NAND and compare the two.
If there was an S/N mismatch, the test would fail.

To accompish these aims, we add an extra NAND location pointer for the
CCA S/N and an extra environment variable for the CCA S/N.

Signed-off-by: Jaeden Amero <jaeden.amero@ni.com>
Acked-by: Joe Hershberger <joe.hershberger@ni.com>