zynq_gem_wrap: Allow skipping PHY reset
Listen to a new environment variable, "ethphyreset", which, when set to
a "no", will skip resetting the PHY. This is useful for configuring the
PHY manually before initializing the networking stack. For example, one
could tell the PHY not to advertise that it is capable of 1000 Mbps
Ethernet, set "ethphyreset" to "no", and then ping or tftp something and
have the manual settings stick and for the link to connect at 100 Mbps.

Signed-off-by: Jaeden Amero <jaeden.amero@ni.com>
Reviewed-by: Joe Hershberger <joe.hershberger@ni.com>