zynq_gem_wrap: Allow skipping PHY renegotiation
Listen to a new environment variable, "ethphyrenegotiate", that, when
set to "no", will skip asking any PHY to renegotiate link. This is
useful for configuring the PHY and bringing up link manually, when used
in conjunction with "ethphyreset". For example, one could manually link
the PHY at whatever speed they wanted and not have to wait for any
auto-negotiation to occur.

Signed-off-by: Jaeden Amero <jaeden.amero@ni.com>
Reviewed-by: Joe Hershberger <joe.hershberger@ni.com>