tftp: Listen to tftpretries
Listen to a new environment variable, "tftpretries", which will control
the number of times a tftp operation will be retried after a timeout.
For instance, if set to 10, the TFTP command will retry up to 10 times
(trying a total of 11 times) to perform the operation. This is useful
when it is desired that the "tftp" command fail quickly (0 retries) or
when it is desired that the "tftp" command should retry many times (1000
retries).

For compatibility with the TftpRRQTimeoutCountMax global used to specify
a custom retry limit just for the initial TFTP connection, we give
priority to the number of retries specified by TftpRRQTimeoutCountMax
over the number of retries specified via "tftpretries". Specifically, if
TftpRRQTimeoutCountMax is changed from its default of TIMEOUT_COUNT, we
will use the TftpRRQTimeoutCountMax value instead of the "tftpretries"
value.

Signed-off-by: Jaeden Amero <jaeden.amero@ni.com>
Reviewed-by: Joe Hershberger <joe.hershberger@ni.com>