nizynq: Make the usb host/device port switching code flexible
The code for USB host/device port switching was hard-coded to handle
USB port 0 for sbRIO-9651. sbRIO-96xx would like to re-use the same
code but for USB port 1. The USB host/device port switching code is
modified to make it flexible enough to support either USB 0 or USB 1.
A target needs to define the base address of the USB port supporting
device/host mode switching in CONFIG_OTG_USB_BASE_ADDR. The code
assumes a standard Zynq USB node in the device tree or it can be
overrided by defining the path to the USB node in the device tree
as CONFIG_OTG_USB_DT_NODE.

Signed-off-by: Keng Soon Cheah <keng.soon.cheah@ni.com>