nicrio: Convert long exclusion logic to positive logic
The NI cRIO product code contains code with long exclusion logic
and make the code less readable. The problem is caused by multiple
products inheriting code that was designed to just handle cRIO-9068
when it was first created. In order to maintain backward
compatibility, exclusion logic is used whenever the support for a new
cRIO product is added. The length of exclusion logic then grows with
the number of new products introduced and eventually make the code
cryptic. An explicit option is now defined for cRIO-9068 and the long
exclusion logic are converted to positive logic to improve code
readability.

Signed-off-by: Keng Soon Cheah <keng.soon.cheah@ni.com>