serial: Add nulldev serial device
Add nulldev as a serial device. This allows the default console to be
specified as the nulldev.  This is specifically helpful when the real
serial console's initialization cannot run early in the boot process.
When the init can be run, then the console can be switched to the real
device.

Signed-off-by: Keng Soon Cheah <keng.soon.cheah@ni.com>