nicrio: Enable MACB Driver Support for Zynq GEM
The NI cRIO product are using PL to convert the GMII to RGMII signal
which this conversion required changing in "tx_clk" clock speed and
also assertion of EMIO signal to indicated different linkspeed.

In the MACB driver, there is no knowledge on the PL GMII to RGMII
convertion. Hence, this change will add in the code for changing the
"tx_clk" and assert EMIO signal when change in linkspeed for the GMII
to RGMII convertion logic to work properly.

Signed-off-by: Wilson Lee <wilson.lee@ni.com>