nicrio: Add RAM size declaration in u-boot
The new u-boot(2016.07) had changes their ram size declaration method.
The previous method which using gd->ram_size to declare the device RAM
size is not longer applicable. This will cause the Linux unable to boot
because of u-boot fail to populate the actual device RAM size. This
changes is to add in the device RAM declaration features into the new
u-boot.

Signed-off-by: Wilson Lee <wilson.lee@ni.com>