dts: zynq: Remove 'dm-pre-reloc' declaration in zynq-7000 dts
In the zynq-7000 devices tree file, there have "u-boot,dm-pre-reloc"
flag declared in the "amba" node that will force all the "amba"
devices under the node to be binded before u-boot reallocate and
before DRAM was avaliable. This will cause the system run out of heap
(OCM), fail to perform malloc and hang. This change is to remove the
"u-boot,dm-pre-reloc" declaration in zynq-7000 dts file.

Signed-off-by: Wilson Lee <wilson.lee@ni.com>