nicriozynq: Change #error to #warning for board name and device code
The compiler will fail the compilation when it is compiling the U-boot
tools code. This failure was cause by no "CONFIG_DEVICE_CODE" and
"CONFIG_NI_BOARD_NAME" supplied during compilation of u-boot tools.

This commit is to change the "#error" to "#warning" to prevent the
compiler failing the compilation when it's compile u-boot tools.

Signed-off-by: Wilson Lee <wilson.lee@ni.com>