lib: sha1: Change uint8_t to unsigned char for sha1_der_prefix array.
uint8_t used in sha1_der_prefix array was not able to recognise by
complier when try to build the tools using 'HOSTCC'. That is because,
in 'HOSTCC' build, the asm/type.h was not included. Hence, the complier
will complain error for uint8_t. This commit is to change the uint8_t
to unsigned char for sha1_der_perfix[] array.

Signed-off-by: Wilson Lee <wilson.lee@ni.com>