ni: Makefile: Enable "arm-gcc" complier to build for u-boot tool.
The fw_printenv and fdtview u-boot tools was consume and used by target
Linux. However in v2017.03 u-boot, the script/basic/fixdep.c will
required host complier to build and execute on the host machine at the
beginning of the build. This lead to switch of complier is needed during
middle of the build process. This commit is to enable the use of host
gcc complier for building the script_basic and allow the use of arm gcc
complier to compile the tools executable(fw_printenv & fdtview) for ARM
target.

Signed-off-by: Wilson Lee <wilson.lee@ni.com>