ni: elvisiii: Enable support of DIO/UART multiplex for shared pins
In ELVIS III target, the DIO/UART have to share pins. Hence, some
machanism needed to notify FPGA either the shared pins is in DIO/UART
mode.

This commit is to enable the use of EMIO 56 to notify PL FPGA either
multiplex to DIO or UART mode for the shared pins in ELVIS III.

Signed-off-by: Wilson Lee <wilson.lee@ni.com>