ni: elvisiii: Rename ELVIS III PLUS and remove ELVIS III
According to marketing feedback, we will not release ELVIS III before
ELVIS III PLUS. Hence, using ELVIS III PLUS naming will be weird to
customer.

This commit is to rename the ELVIS III PLUS to ELVIS III and remove the
ELVIS III target as it will not release at this moment.

Signed-off-by: Wilson Lee <wilson.lee@ni.com>
Acked-by: Keng Soon Cheah <keng.soon.cheah@ni.com>
Acked-by: Joseph Hershberger <joseph.hershberger@ni.com>
Natinst-Reviewboard-ID: 214414