nicrio: Add wireless region and board id to backup page
U-Boot will restore wireless region and board id from backup page if they are
missing. The default value for wireless region and board id is also removed to
not assume any value.

The wireless region and wireless board id are stored as Pascal string in backup
page and 5 bytes are allocated for each of them. These values are populated
when the device is manufactured. Length of string should be explicitly set to 0
if this information is not applicable.

Signed-off-by: Keng Soon Cheah <keng.soon.cheah@ni.com>