ubifs: fix stale superblock pointer after ubifs is unmounted
The ubifs superblock is added to the file system superblock linked-list when it
is mounted, via hlist_add_head(&s->s_instances, &type->fs_supers). However, the
superblock node is not removed from the linked-link when it is umounted and
freed. This is causing the superblock linked-list carrying a stale pointer to
freed memory region and subsequently causes issue when it is referenced again.

Fix this problem by removing the superblock node from linked-list before the
superblock is freed.

Signed-off-by: Keng Soon Cheah <keng.soon.cheah@ni.com>