nimyrio: Enable MACB Driver Support for Zynq GEM
The NI product are using PL to convert the GMII to RGMII signal
which this conversion required changing in "tx_clk" clock speed and
also assertion of EMIO signal to indicated different linkspeed.

In the MACB driver, there is no knowledge on the PL GMII to RGMII
convertion. Hence, this change will add in the code for changing the
"tx_clk" and assert EMIO signal when change in linkspeed for the GMII
to RGMII convertion logic to work properly.

Integrate from commit: 4078f9624a1bc1344ebc1ad713683afa141e2086

Signed-off-by: Kim Khiam Lim <kim.khiam.lim@ni.com>
Reviewed-by: Keng Soon Cheah <keng.soon.cheah@ni.com>
---
Natinst-Reviewboard-ID: 353955